Second commit

+++ Задание к уроку Основы Ruby. Часть 1/area_of_a_riangle.rb
@@ -1,7 +1,7 @@
 print "Введите основание треугольника"
-length = gets.chomp.to_i
+length = gets.chomp.to_f
 
 print "Введите высоту треугольника"
-height = gets.chomp.to_i
+height = gets.chomp.to_f
 
 puts "#{0.5 * length * height}"
